Add property to determine whether to use shadow copying; fix bug in creation of AppDomain when appbase has been specified.

+++ src/ClientUtilities/util/TestDomain.cs
@@ -75,9 +75,19 @@ namespace NUnit.Util
 		/// </summary>
 		private EventListener listener;
 
+		/// <summary>
+		/// Indicate whether files should be shadow copied
+		/// </summary>
+		private bool shadowCopyFiles = true;
+
 		#endregion
 
 		#region Properties
+
+		public AppDomain AppDomain
+		{
+			get { return domain; }
+		}
 
 		public TextWriter Out
 		{
@@ -126,6 +136,17 @@ namespace NUnit.Util
 		public Version FrameworkVersion
 		{
 			get { return Runner.FrameworkVersion; }
+		}
+
+		public bool ShadowCopyFiles
+		{
+			get { return shadowCopyFiles; }
+			set
+			{
+				if ( this.domain != null )
+					throw new ArgumentException( "ShadowCopyFiles may not be set after domain is created" );
+				shadowCopyFiles = value;
+			}
 		}
 
 		#endregion
@@ -203,18 +224,17 @@ namespace NUnit.Util
 
 		public Test Load( NUnitProject project )
 		{
-			if ( project.IsAssemblyWrapper )
-				return Load( project.ActiveConfig.Assemblies[0].FullPath );
-			else
-				return Load( project.ProjectPath, project.ActiveConfig.TestAssemblies );
+			return Load( project, null );
 		}
 
 		public Test Load( NUnitProject project, string testFixture )
 		{
+			ProjectConfig cfg = project.ActiveConfig;
+
 			if ( project.IsAssemblyWrapper )
-				return Load( project.ActiveConfig.Assemblies[0].FullPath, testFixture );
+				return Load( cfg.Assemblies[0].FullPath, testFixture );
 			else
-				return Load( project.ProjectPath, project.ActiveConfig.TestAssemblies, testFixture );
+				return Load( project.ProjectPath, cfg.BasePath, cfg.ConfigurationFile, cfg.PrivateBinPath, cfg.TestAssemblies, testFixture );
 		}
 
 		public void Unload()
@@ -226,8 +246,11 @@ namespace NUnit.Util
 				try
 				{
 					AppDomain.Unload(domain);
-					DirectoryInfo cacheDir = new DirectoryInfo(cachePath);
-					if(cacheDir.Exists) cacheDir.Delete(true);
+					if ( this.ShadowCopyFiles )
+					{
+						DirectoryInfo cacheDir = new DirectoryInfo(cachePath);
+						if(cacheDir.Exists) cacheDir.Delete(true);
+					}
 				}
 				catch( CannotUnloadAppDomainException )
 				{
@@ -401,16 +424,6 @@ namespace NUnit.Util
 			domain = MakeAppDomain( testFileName, appBase, configFile, binPath );
 		}
 
-		private void CreateDomain( NUnitProject project )
-		{
-			ProjectConfig cfg = project.ActiveConfig;
-
-			if ( project.IsAssemblyWrapper )
-				CreateDomain( cfg.Assemblies[0].FullPath );
-			else
-				CreateDomain( project.ProjectPath, cfg.BasePath, cfg.ConfigurationFilePath, cfg.PrivateBinPath, cfg.TestAssemblies );
-		}
-
 		/// <summary>
 		/// This method creates appDomains for the framework.
 		/// </summary>
@@ -428,11 +441,18 @@ namespace NUnit.Util
 
 			// We always use the same application name
 			setup.ApplicationName = "Tests";
-			// We always want to do shadow copying. Note that we do NOT
+			// Note that we do NOT
 			// set ShadowCopyDirectories because we  rely on the default
 			// setting of ApplicationBase plus PrivateBinPath
-			setup.ShadowCopyFiles = "true";
-			setup.ShadowCopyDirectories = appBase;
+			if ( this.ShadowCopyFiles )
+			{
+				setup.ShadowCopyFiles = "true";
+				setup.ShadowCopyDirectories = appBase;
+			}
+			else
+			{
+				setup.ShadowCopyFiles = "false";
+			}
 
 			setup.ApplicationBase = appBase;
 			setup.ConfigurationFile =  configFile;
@@ -440,7 +460,8 @@ namespace NUnit.Util
 
 			AppDomain runnerDomain = AppDomain.CreateDomain(domainName, evidence, setup);
 			
-			ConfigureCachePath(runnerDomain);
+			if ( this.ShadowCopyFiles )
+				ConfigureCachePath(runnerDomain);
 
 			return runnerDomain;
 		}
